requirements.txt
certifi==2022.12.7
chardet==5.1.0
Django==4.1.4
django-environ==0.9.0
django-filter==22.1
django-jazzmin==2.6.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-utils-six==2.0
et-xmlfile==1.1.0
idna==3.4
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
Pillow==9.3.0
postgis==1.0.4
psycopg-postgis==0.4.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
